\name{missing.total}
\alias{missing.total}
\title{A function to find the Number of missing elements.}
\description{
A function which will return the number of missing items for a data frame with just 1 column.
}
\usage{
missing.total(x)
}

\arguments{
  \item{x}{A data frame with a single column}
}
\details{}
\value{The number of missing values in the data frame}
\references{}
\author{Nishan Mudalige}
\note{}
\seealso{}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
